seperated optimizelyConfigService tests
